+++ integration-tests/src/tests/client/features.rs
@@ -1,2 +1,4 @@
+#[cfg(feature = "protocol_feature_account_id_in_function_call_permission")]
+mod account_id_in_function_call_permission;
 #[cfg(feature = "protocol_feature_fix_contract_loading_cost")]
 mod fix_contract_loading_cost;
